Migrate reshaping ops
