+++ sharktank/sharktank/ops/shaping/__init__.py
@@ -1,0 +1,22 @@
+# Shaping operations
+from .reshape import reshape
+from .view import view
+from .flatten import flatten
+from .unflatten import unflatten
+from .transpose import transpose
+from .permute import permute
+from .squeeze import squeeze
+from .unsqueeze import unsqueeze
+from .expand import expand
+
+__all__ = [
+    "reshape",
+    "view",
+    "flatten",
+    "unflatten",
+    "transpose",
+    "permute",
+    "squeeze",
+    "unsqueeze",
+    "expand",
+]
